requirements.txt
blinker==1.9.0
boto3==1.39.2
botocore==1.39.2
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
mysql-connector==2.2.9
mysql-connector-python==9.3.0
PyJWT==2.10.1
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
